Moved routes settings initialization and configuration into frontned
refs #9528

- The routes.yaml settings was designed to handle more settings then just routing, but till this day there wasn't anything else added. As routes.yaml is only being used by frontend router it seems like a logical move to have this code in frontend so that it doesn't have to reach out to server.
