+++ core/frontend/services/routing/bootstrap.js
@@ -1,7 +1,7 @@
 const debug = require('ghost-ignition').debug('services:routing:bootstrap');
 const _ = require('lodash');
 const common = require('../../../server/lib/common');
-const settingsService = require('../../../server/services/settings');
+const settingsService = require('../settings');
 const themeService = require('../themes');
 const StaticRoutesRouter = require('./StaticRoutesRouter');
 const StaticPagesRouter = require('./StaticPagesRouter');
