+++ core/frontend/services/settings/default-routes.yaml
