+++ core/frontend/services/settings/yaml-parser.js
@@ -1,6 +1,6 @@
 const yaml = require('js-yaml'),
     debug = require('ghost-ignition').debug('services:settings:yaml-parser'),
-    common = require('../../lib/common');
+    common = require('../../../server/lib/common');
 
 /**
  * Takes a YAML file, parses it and returns a JSON Object
