+++ core/server/config/overrides.json
@@ -6,7 +6,7 @@
         "helperTemplates": "core/frontend/helpers/tpl/",
         "adminViews": "core/server/web/admin/views/",
         "defaultViews": "core/server/views/",
-        "defaultSettings": "core/server/services/settings/",
+        "defaultSettings": "core/frontend/services/settings/",
         "internalAppPath": "core/frontend/apps/",
         "internalStoragePath": "core/server/adapters/storage/",
         "internalSchedulingPath": "core/server/adapters/scheduling/",
