+++ core/test/unit/services/settings/validate_spec.js
@@ -2,7 +2,7 @@ const should = require('should');
 const sinon = require('sinon');
 const common = require('../../../../server/lib/common');
 const themesService = require('../../../../frontend/services/themes');
-const validate = require('../../../../server/services/settings/validate');
+const validate = require('../../../../frontend/services/settings/validate');
 
 should.equal(true, true);
